+++ src/module/__test__/cache/packageCache.test.ts
@@ -28,6 +28,7 @@ const packageSettings: ParsedPackageSettings = {
   externallyImported: [],
   ignorePatterns: [],
   ignoreOverridePatterns: [],
+  defaultIgnoreOverrides: [],
   testFilePatterns: [],
 };